+++ lua/sql-lens/connections/base.lua
@@ -31,4 +31,22 @@ function Base:lint(sql, cb)
   cb(nil)
 end
 
+---@param cb fun(err?: string, tables?: string[])
+function Base:list_tables(cb)
+  cb("Not implemented", {})
+end
+
+---@param tbl string
+---@param cb fun(err?: string, columns?: string[])
+function Base:list_columns(tbl, cb)
+  cb("Not implemented", {})
+end
+
+---Return a preview query for the given table
+---@param tbl string
+---@return string
+function Base:preview_query(tbl)
+  return "SELECT * FROM " .. tbl .. " LIMIT 50;"
+end
+
 return Base